final fixes

+++ DataStructure/LinkedListDS/NurseFunctions.cpp
@@ -291,11 +291,11 @@ void printNursesForShift(const ShiftSchedule& schedule, int shift) {
  * with the highest score along with the score value. It also prints the associated schedule.
  */
 void returnBestSatisfactionScores(
-    double scoreGeneticAlgorithm, const ShiftSchedule& scheduleGeneticAlgorithm,
-    double scoreBruteForce, const ShiftSchedule& scheduleBruteForce,
-    double scoreLinearProgramming, const ShiftSchedule& scheduleLinearProgramming
+    int scoreGeneticAlgorithm, const ShiftSchedule& scheduleGeneticAlgorithm,
+    int scoreBruteForce, const ShiftSchedule& scheduleBruteForce,
+    int scoreLinearProgramming, const ShiftSchedule& scheduleLinearProgramming
 ) {
-    double highestScore = scoreGeneticAlgorithm;
+    int highestScore = scoreGeneticAlgorithm;
     std::string bestMethod = "Genetic Algorithm";
     const ShiftSchedule* bestSchedule = &scheduleGeneticAlgorithm; // Use a pointer to avoid reassignment issues
